Wrote tests for features below:
Wrote tests for features below:

Added the ability to specify the foreign key column on the current
(base) table (i.e. the table for the model in which the has_one or
has_many relationship method is being defined). This applies to
_has_one_or_many(...), has_one(...) and has_many(...)

Added the ability to specify the foreign key column in the associated
(related) table for the belongs_to relationship. This applies to
belongs_to(...)

Added the ability to specify the foreign key column in the base table
and the foreign key column in the associated (related) table for a
has_many_through relationship. This applies to has_many_through(...)