Confusion between related and base table
Please review edit. I believe there is a mistake in the description of Has One.  The original statement is very confusing until you realize that there is a mistake.  It is sort of like saying: The base table ( in this case user) references the foreign key in the base table ( in this case user).